+++ lib/mocker/src/replay/online/mod.rs
@@ -1,9 +1,15 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-mod runtime;
+mod demux;
+mod live_runtime;
+mod state;
+mod task;
 
-pub(crate) use runtime::{
+#[cfg(test)]
+mod tests;
+
+pub(crate) use live_runtime::{
     simulate_concurrency_requests, simulate_concurrency_workload, simulate_trace_requests,
     simulate_trace_workload,
 };
